Remove non raylet code.
