+++ java/ray.config.ini
@@ -44,8 +44,6 @@ driver_id = 0123456789abcdef0123456789abcdef01234567
 
 redis_port = 34111
 
-num_local_schedulers = 1
-
 max_submit_task_buffer_size_bytes = 51200
 
 default_first_check_timeout_ms = 1000
@@ -59,8 +57,6 @@ object_store_occupied_memory_MB = 2
 deploy = false
 
 onebox_delay_seconds_before_run_app_logic = 0
-
-use_raylet = true
 
 static_resources = CPU:4,GPU:0
 
@@ -122,9 +118,6 @@ driver_args =
 redis_server = %CONFIG_FILE_DIR%/../build/src/common/thirdparty/redis/src/redis-server
 redis_module = %CONFIG_FILE_DIR%/../build/src/common/redis_module/libray_redis_module.so
 store = %CONFIG_FILE_DIR%/../build/src/plasma/plasma_store_server
-store_manager = %CONFIG_FILE_DIR%/../build/src/plasma/plasma_manager
-local_scheduler = %CONFIG_FILE_DIR%/../build/src/local_scheduler/local_scheduler
-global_scheduler = %CONFIG_FILE_DIR%/../build/src/global_scheduler/global_scheduler
 raylet = %CONFIG_FILE_DIR%/../build/src/ray/raylet/raylet
 python_dir = %CONFIG_FILE_DIR%/../build/
 java_runtime_rewritten_jars_dir =
@@ -135,9 +128,6 @@ java_jnilib_paths = ray.java.path.jni.package
 redis_server = %CONFIG_FILE_DIR%/../build/src/common/thirdparty/redis/src/redis-server
 redis_module = %CONFIG_FILE_DIR%/../build/src/common/redis_module/libray_redis_module.so
 store = %CONFIG_FILE_DIR%/../build/src/plasma/plasma_store_server
-store_manager = %CONFIG_FILE_DIR%/../build/src/plasma/plasma_manager
-local_scheduler = %CONFIG_FILE_DIR%/../build/src/local_scheduler/local_scheduler
-global_scheduler = %CONFIG_FILE_DIR%/../build/src/global_scheduler/global_scheduler
 raylet = %CONFIG_FILE_DIR%/../build/src/ray/raylet/raylet
 python_dir = %CONFIG_FILE_DIR%/../build/
 java_runtime_rewritten_jars_dir =
@@ -148,9 +138,6 @@ java_jnilib_paths = ray.java.path.jni.package
 redis_server = %CONFIG_FILE_DIR%/native/bin/redis-server
 redis_module = %CONFIG_FILE_DIR%/native/lib/libray_redis_module.so
 store = %CONFIG_FILE_DIR%/native/bin/plasma_store_server
-store_manager = %CONFIG_FILE_DIR%/native/bin/plasma_manager
-local_scheduler = %CONFIG_FILE_DIR%/native/bin/local_scheduler
-global_scheduler = %CONFIG_FILE_DIR%/native/bin/global_scheduler
 raylet = %CONFIG_FILE_DIR%/native/bin/raylet
 python_dir = %CONFIG_FILE_DIR%/python
 java_runtime_rewritten_jars_dir = %CONFIG_FILE_DIR%/java/lib/
