+++ java/test/src/main/java/org/ray/api/test/ResourcesManagementTest.java
@@ -49,7 +49,6 @@ public class ResourcesManagementTest {
 
   @Test
   public void testMethods() {
-    Assume.assumeTrue(AbstractRayRuntime.getParams().use_raylet);
     // This is a case that can satisfy required resources.
     RayObject<Integer> result1 = Ray.call(ResourcesManagementTest::echo1, 100);
     Assert.assertEquals(100, (int) result1.get());
@@ -64,7 +63,6 @@ public class ResourcesManagementTest {
 
   @Test
   public void testActors() {
-    Assume.assumeTrue(AbstractRayRuntime.getParams().use_raylet);
     // This is a case that can satisfy required resources.
     RayActor<ResourcesManagementTest.Echo1> echo1 = Ray.createActor(Echo1::new);
     final RayObject<Integer> result1 = Ray.call(Echo1::echo, echo1, 100);
